Refactoring of TracCrossReferences: create `trac.xref` module
The cross-referencing facility has been separated from the TracObject.

In the `trac.xref` module, there's now a separate file for the Web UI
and the macros (that's fine, as there will be more macros in the future).

Finally, the API for the xref has been simplified and the documentation
has been updated.