VcRefactoring: reintroduce non-greedy match for InterTrac prefix in shorthand links.
This fixes the change made in [2872], when support for hex digits in changesets was added. Note that I got the comment wrong there, as that changeset actually made the match ''greedy''.

The problem with greediness is that changesets starting with hex digits can't be recognized at all (e.g. in `[b47]`, "b" will be the intertrac prefix).

Therefore, I think it's better to drop support for links like `[trac2781]` that motivated the original change, as there are alternatives: either `[trac 2781]` or `[T2781]` if "T" is set as an alias for "trac".