Another attempt at fixing the unicode problem with the reports.
Actually, the reason for the trouble was that pysqlite2 
returns the column description information as an UTF-8 
encoded string.

Also, I reverted r3084 and instead do the conversion
in `HDFWrapper.add_value` in a "robust" and backward
compatible way: if a plugin store UTF-8 strings in the
HDF, things will work as expected. If the string was
not encoded using UTF-8, replacement characters will be
inserted.

This should really close #2905 this time :)