More robust parsing and formatting of Wiki lists (#124)
Previously, each indent of 2 spaces would introduce a new indentation level.
This lead to a confusing output in a variety of situations, 
especially in changelog style commit messages.

Now, new indent levels are only introduced when they need to be,
i.e. when there's actually a new indented list or paragraph:

{{{
     * first level
           * second level
             - third level

  First level
                   Second level
}}}

Note that a tab character is converted to 8 space characters 
before computing the indentation level.

In addition, lists can now also be started using "-" (as this is commonly found
 in changelogs as well).

There will be a follow-up changeset focused on the interaction between lists and quotes.