Make it possible to delete the latest revision of a Wiki page directly from the diff view.
Example: when undesirable content is seen in a wiki page, it's now simply a matter of going to the ''Last Change'' view, and from there, delete this version.

Another possibility is to delete a ''range'' of versions, when examining such a range in the diff view.
However, this is only possible if that range of versions extends to the latest revision, or if there's no net change introduced by this range of versions (e.g. spam followed by spam removal).

Closes #1114.