Improved internal error reporting:
Internal errors are eith indicative of a problem with the Trac installation or a bug in Trac itself.
In both cases, the Trac admnistrator must be the first informed, as he's the most adequate person to either fix the installation issue or to relay the bug report to the TracTeam.

This is why upon internal errors, "normal" users will now be told to contact their local Trac administrator.
A new `[project]` setting, ''admin'', can be used to define a mail contact. If this contact information is not defined, a button for creating a ticket on the local Trac will added instead of the mailto: link.

When someone with the TRAC_ADMIN permission hits an internal error, however, he will have the full stacktrace information as before, plus some system information obtained from the new `self.env.systeminfo` dictionary.
In addition, he will be advised to ''search'' on Trac's Trac for a similar issue, and will have the possibility to create a ticket directly on t.e.o, with all the information we usually ask for.

The ''Create'' button will only prepare a new ticket (''preview'' mode). That way, the description can be further edited and other ticket fields can be set before submitting.


''This change could eventually be backported to [milestone:0.10.1]''