VcRefactoring/BrowserQuickjump: implementing cmlenz's suggestion of merging the `branches` and `branch_folders` configuration options (and the same for tag)
Now there's only `branches` and `tags`:
{{{
[svn]
branches = trunk,branches/*
tags = tags/*
}}}
i.e. a "*" is used to specify that the content of the path should be used, instead of the path itself (and only subfolders are taken into account, not files).

Not only this simplifies the code, but this allows you to freely mix the fixed paths and the dynamic paths, in the order you want.