Blame branch: when clicking on a blame column, the changeset view is dynamically fetched (XHR).
This changeset view is displayed in a separate panel, on top of the blame view. It also displays the diffs concerning the file being annotated. This make it possible to better grasp the context of the change (e.g. by seeing the deleted lines).

Also, the browser links in the diff lines are set to remain in ''annotate'' mode. This makes it possible to conveniently dig the history of a specific part of the code.

Finally, the highlighting of all the rows of a given revision is only done when clicking on a blame column, not anymore when hovering on the blame column.
Not only was this too slow for big files, but doing so makes it clear which are the lines in the file that are belonging to the changeset being displayed.