Blame branch: better handle failure of an annotator
Instead of retrying all preview renderers and failing again for each of them, an error with an annotator (assuming it happens at the `get_annotation_data` stage) will now simply omit the result of this annotation from the output, leaving the error message accessible in the title info.

In the case of the blame annotator, such error typically correspond to SVN thinking it can't get the annotations because the file is binary. With `client.blame3`, it would be possible to override this check, but only `client.blame2` is currently usable in the bindings as of Subversion 1.4.2.