Property renderer branch: add IPropertyRenderer and IPropertyDiffRenderer interfaces
This also provides default implementation for those interfaces, which reproduce the current behavior, except for the default `DefaultPropertyDiffRenderer` which implements #3002 (view difference of multi-line properties as rendered diffs).

In order to implement the above, a few changes in the `PatchRenderer` were needed.

Also, the `.revision` property of the WikiContext which was added in the blame branch was also needed here, so I ported that part of r4483.

The main ticket for this is #1601.