Property renderer branch: reimplement #2545 to use IPropertyRenderer
This modifies the return type of `Changeset.get_properties` in a backward incompatible way.

The `mode` used in IPropertyRenderer for rendering revision properties is `'revprop'`.

In order to make it easier to quickly get some useful rendererings, I've added a `WikiPropertyRenderer`, which learn from the TracIni file what properties could be rendered as wiki text:
{{{
[browser]
wiki_properties = trac:description
oneliner_properties = trac:summary
}}}
(note that the above correspond to the defaults).

Besides, a few additional changes in the templates have been made so that the Changeset view now validates.