Timeline link resolver branch (#975):
This implements a `timeline:` link resolver prefix.
The argument must be a date string in the ISO:8601 format, with at least the year and the month specified.

Besides TracLinks, there are many more ways to go back to the timeline, now: I modified the format_datetime helper function and the dateinfo macros so that we get links to the timeline from various places (e.g. changeset time, ticket created and last modified time, browser age, wiki history, ...).

Also, there are some fine tuning to the datetime utilities in `trac.util.datefmt`:
 - the `format_*` functions supports an `'iso8601'` format
 - `parse_date` can also understand this format
 - fixed offsets timezones are now using strings like "GMT +01:00" instead of the more confusing 'Etc/GMT-1' one (see discussion in #975).