Property renderer branch: allow property renderers to gain more control over the rendering
If the text or markup output implemented so far doesn't provide enough control, one can now return a `RenderedProperty` instance. This disables the default blurb associated to properties (like ''Property '''blah''' set to ...'') and allows to specify the full `content`. In the case of #2545 (revprops) the `name` can also be set, as well as the attributes for the <dt> and <dl> elements containing the name and the content, respectively.