PermissionPolicy: changed the semantic of `perm.require(*actions)`
Now, the check passes if ''any'' of the listed actions is permitted.
This is because a `require` check for ''all'' actions can simply be done by a succession of `require()` calls.

The above is used for the `Attachment._render_list` check, where a LIST permission can also be used instead of the VIEW one in order to get access to the list of attachments (addresses #4240).

Also, the `PermissionCache.require` is now compatible again with old-style `req.perm.assert_permission('...')` checks.

Finally, the message corresponding to a `PermissionError` has been made more explicit.