Add the possibility to save a custom query as a report (#1735)
Reports can now contain a query string, either in the '?...' URL form or in the TracQuery#QueryLanguage form.
Viewing such a report simply redirects to the corresponding custom query. 

When a saved custom query is displayed, the title and the description of the report will also be shown, as well as an ''Edit Query'' button, for editing the query as it was saved.

There will also be a ''Save Query'' button, which will save the ''current'' parameters of the query, either in the corresponding report if there's already one or in a new report, provided the appropriate permissions are satisfied (REPORT_MODIFY and REPORT_CREATE, respectively).