WorkFlow: Interface and basic implementation of ITicketActionController. The basic implementation pulls from the config or uses defaults if not present in the config.
web_ui is using this to present action options and perform
state transitions. Assign to needs to be tested more,
but it uses the same operations as before. All other
operations are clearly doing 'the right thing'.

Currently, any conflict in opinion about state transitions
between ticket action controllers results in a TracError.