VcRefactoring/Controller: simplify the IRequestHandler implementation in case a !RequestProcessor is used.
The !RequestProcessor has now a `match` class method which makes use of the `url_pattern` property of the subclass in order to match and parse the `req.path_info`. The match method returns a new `RequestProcessor` instance on success, `None` otherwise.

That new instance will then be used directly later on to "process" the request.