VcRefactoring/Controller: introduced a Controller mixin, which takes care implementing the IRequestHandler methods (actually only match_request is needed in the new scheme).
The Component inheriting from Controller must simply set the `request_processor` property to the name of the appropriate `RequestProcessor` subclass.

This has the unfortunate consequence that the subclass must be known at the time this property is set, requiring to move all the already defined `RequestProcessor` classes to the top of their files. 

Maybe separate files would be more appropriate?