TracTimeline API: in 0.10, the changeset events in RSS where including a summary of the commit message.
After the initial refactoring of the Timeline API in 0.11dev, there was only the changeset number left, and this was not acceptable.

So while looking for a somewhat clean solution to the issue, I ended up modifying the !TimelineEvent class a lot. Instead of a fixed "markup" and "wikitext" field, one can associate multiple named markup and wikitext fragments to the event. This is more flexible and allows to build different styles of timelines.

For example, the standard HTML timeline shows for each event its markup 'title', then the markup 'header', followed by the wikitext 'body', followed by the markup 'footer'.
The standard RSS timeline uses the markup 'summary' if present in preference to the 'title', for its <title> element.

For determining the way the wiki formatting of a wikitext is done, the event's provider is queried using the new `event_formatter` interface method.

Finally, the `href` corresponding to the event should in the end be the one of its associated resource descriptor (context for now).