Change Wiki navigation links from:
{{{
Start Page | Index by Title | Index by Date | Last Change
}}}

to:

{{{
Start Page | Index | History | Last Change
}}}

This enables direct access to both the revision log of a page and the diff corresponding to the last modification, see #5491.

A `trac-admin ... wiki upgrade` is needed, as the RecentChanges and TitleIndex pages now need to link to each other. The ''Index'' navigation link goes to the latter.