Improved rendering of ticket property changes (merged from the FieldRefactoring branch).
The rendering of property change is now done in `rendered_changelog_entries` in the controller rather than in the template. That method can therefore be called by `_insert_ticket_data` (for the normal ticket display) and by `_get_history` (displaying ticket version changes). In the future, that method will delegate the rendering to dedicated field renderer classes and/or property renderer components.

The `rendered_changelog_entries` method implements a few additional rendering enhancements:
 - improved visualization of changes to the Ticket CC list (added/removed and obfuscation of e-mail addresses if needed) (fixes #5126)
 - similar added/removed rendering for changes to the keyword list
 - set/unset for checkbox custom fields