Don't try to figure out the exact oldest rev in case of scoped Subversion revs.
We don't do that because it has apparently a big perfomance impact, and moreover is the trigger for a huge memory leak, as documented in #5213.

However, getting the exact start of a the history in a scoped repository was important for the resync, so we still have to do that at this occasion.

I keep the lines as comments instead of deleting them completely, as a reminder to investigate the reasons of that leak.

Closes #5213.