Attachment metadata can be searched.
In a way similar to what is done for the `ITimelineEventProvider`, the Attachment module provides a `get_search_results(self, req, resource_realm, terms)` helper that can be used by the various modules providing attachment support. The search for attachment therefore becomes dependent on the selection of the search source for the parent resource.

Patch by Remy Blank, closes #2561.