rework-testing branch: fix `FunctionalTestEnvironment.destroy` on Windows, which wasn't able to remove the test environment due to the newly introduced get_trac_environment call:  - the db/trac.db was still opened for reading  - the log/trac.log was still opened for writing
Doing a `env.shutdown()` was enough to get rid of the first problem, but for the second, the log handler had also to be cleaned-up, so shutdown was modified to that effect.

There's apparently still one user of the logger ''after'' the environment has been destroyed, as we see the following:
{{{
...
Test attachment: raw format as explicit argument ... ok
No handlers could be found for logger "...\repos\rework-testing\testenv\trac"
Test for regression of the plugin reload fix in r6017 ... ok
...
}}}
but so far I couldn't locate it (not ''that'' important either).