Follow-up to the merge of the expanded unicode support in Config (r8469).
This broke the creation of query links in ticket view (#7562), as keyword arguments can't be unicode objects in Python 2.x. So we use the already existing feature of Href, which can be given a dict of parameters in its last positional argument.

Also checked other places that could have similar issues.

Completes #8276 on trunk.