TracMercurial for 0.11: support Mercurial 1.4dev (after hg:1ef630452e0b).
Thanks to Benoit Boissinot on #mercurial for the hint about not supporting unreleased Mercurial versions.
Actually an unreleased version will always be considered to be the newest. If this assumption is wrong, upgrading to mpm's hg tip should do, or at least getting a release version so that the coarse grained version check will work.