MultiRepos: make it easier to define an alias to the default legacy repository from the .ini.
i.e. if one has defined the `[trac] repository_dir`, all it takes to make an alias ''def'' for it is to do:
{{{
[repositories]
def =
}}}

Note that this also allows to add properties to the default legacy repository in the .ini more easily, e.g.
{{{
[repositories]
.hidden = true
}}}
(before one needed to add an explicit and redundant `.dir = <path/to/repos>` entry to that section)