error reporting: while testing #8790, an error occurred in  `_('Error')`.
The error during `gettext` was expected, however if anything goes wrong at this level, the consequences are rather severe, as no response is sent (not even an internal error). So we should rather catch and ignore secondary errors, which are anyway likely the same or triggered by the primary (critical) error.