core: simplify `Component` creation.
This replaces the `Component.__new__` method with a `ComponentMeta.__call__` method. The latter calls directly the `Component` initializer, so we avoid synthetize one in `ComponentMeta.__new__` which becomes considerably simpler.

Note that we follow the same approach as taken in r10242, i.e. the newly created `Component` is only registered after its initialization. So we still have a possible race condition by which two different Component instances will be created for the same class but only one will be registered for later reuse. As noted on #8658, this is not that problematic and much better than risking to use an uninitialized Component.

Patch contributed by Jun Omae on #9418.